--- src/test/java/opanas/Calculation.java
@@ -1,16 +1,0 @@
-package opanas;
-
-public class Calculation {
-    public int calculate(int n)
-    {
-        int p=1;//x^0
-        int r=0;
-        do {
-            r++;
-            p=(int)Math.pow(2,r);
-
-        }
-        while (p<=n);
-        return p;
-    }
-}
